# src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: Artigo
  description: Controle de orçamento
  version: 1.0.0
servers:
  - url: /v1
paths:
  /receitas:
    get:
      summary: Recuperar receitas
      operationId: getReceitas
      tags:
        - Receitas
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceitasResponse'
    post:
      summary: Inserir uma receita
      operationId: addReceita
      tags:
        - Receitas
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Receita'
        required: true
      responses:
        201:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receita'

  /receitas/{id}:
    get:
      summary: Recuperar receita
      operationId: getReceita
      tags:
        - Receitas
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receita'

    put:
      summary: Atualizar receita
      operationId: AtualizarReceita
      tags:
        - Receitas
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Receita'
      responses:
        202:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receita'

    delete:
      summary: Deletar receita
      operationId: DeleteReceita
      tags:
        - Receitas
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        202:
          description: sucesso

  /receitas/{mes}/{ano}:
    get:
      summary: Recuperar receitas do mês
      operationId: getReceitasMensal
      tags:
        - Receitas
      parameters:
        - $ref: '#/components/parameters/mes'
        - $ref: '#/components/parameters/ano'
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceitasResponse'

  /despesas:
    get:
      summary: Recuperar despesas
      operationId: getDespesas
      tags:
        - Despesas
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DespesasResponse'
    post:
      summary: Inserir uma receita
      operationId: addDespesa
      tags:
        - Despesas
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Despesa'
        required: true
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Despesa'

  /despesas/{id}:
    get:
      summary: Recuperar despesa
      operationId: getDespesa
      tags:
        - Despesas
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Despesa'

    put:
      summary: Atualizar despesa
      operationId: AtualizarDespesa
      tags:
        - Despesas
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Despesa'
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Despesa'

    delete:
      summary: Deletar despesa
      operationId: DeleteDespesa
      tags:
        - Despesas
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        202:
          description: sucesso

  /despesas/{mes}/{ano}:
    get:
      summary: Recuperar despesas do mês
      operationId: getDespesasMensal
      tags:
        - Receitas
      parameters:
        - $ref: '#/components/parameters/mes'
        - $ref: '#/components/parameters/ano'
      responses:
        200:
          description: sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DespesasResponse'

components:
  parameters:
    id:
      in: path
      name: id
      required: true
      schema:
        type: string

    mes:
      in: path
      name: mes
      required: true
      schema:
        type: integer

    ano:
      in: path
      name: ano
      required: true
      schema:
        type: integer

  schemas:

    ReceitasResponse:
      type: object
      properties:
        receitas:
          type: array
          items:
            $ref: '#/components/schemas/Receita'

    Receita:
      type: object
      properties:
        id:
          type: string
          format: int64
        descricao:
          type: string
          example: Descricao da receita
        valor:
          type: number
        data:
          type: string
          format: 'date-time'


    DespesasResponse:
      type: object
      properties:
        despesas:
          type: array
          items:
            $ref: '#/components/schemas/Despesa'

    Despesa:
      type: object
      properties:
        id:
          type: string
          format: int64
        descricao:
          type: string
          example: Descricao da despesa
        valor:
          type: number
        data:
          type: string
          format: 'date-time'
        categoria:
          $ref: '#/components/schemas/Categoria'

    Categoria:
      type: string
      x-enum-varnames: [ ALIMENTACAO, SAUDE, MORADIA, TRANSPORTE, EDUCACAO, LAZER, IMPREVISTOS, OUTRAS ]
      enum: [ "Alimentação", "Saúde", "Moradia", "Transporte", "Educação", " Lazer", "Imprevistos", "Outras" ]